requirements.txt
codeclimate-test-reporter==0.2.3
coverage==5.5
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
flake8==6.1.0
psycopg2==2.9.7
psycopg2-binary==2.9.8
python-dotenv==1.0.0
urllib3==2.0.5
